ReplayEngine now handles its own events rather than requeueing them

+++ HelloProcessing/src/common/timelines/Timeline.java
@@ -3,13 +3,13 @@ package common.timelines;
 public class Timeline {
 	
 	static final long REALTIME_CONVERSION_RATE = 1000000;//converts to milliseconds
-	int ticksize = 1;
+	double ticksize = 1;
 	
 	public long getTime(){
 		if(anchor == null)
-			return (System.nanoTime()/REALTIME_CONVERSION_RATE-origin)/ticksize;
+			return (int)((System.nanoTime()/REALTIME_CONVERSION_RATE-origin)/ticksize);
 		else
-			return (anchor.getTime() - origin)/ticksize;
+			return (int)((anchor.getTime() - origin)/ticksize);
 	}
 	
 	Timeline anchor;
@@ -28,12 +28,12 @@ public class Timeline {
 		this(null, 1);
 	}
 	
-	public void setTicksize(int ticksize){
+	public void setTicksize(double ticksize){
 		this.ticksize = ticksize>0?ticksize:1;
 		System.out.println("Current Ticksize is "+this.ticksize);
 	}
 
-	public int getTicksize() {
+	public double getTicksize() {
 		return this.ticksize;
 	}
 }
